requirements.txt
fastapi
uvicorn[standard]
openai
numpy
scikit-learn
python-dotenv
python-multipart
